configured react navigation and setup initial frontend strucure
