+++ client/components/Homescreen.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import { Text, View } from 'react-native';
+
+export default function Homescreen() {
+  return (
+    <View>
+      <Text>Homescreen</Text>
+    </View>
+  );
+}
